Small refactoring of Manage Lookup

+++ src/Warranty.Core/Features/ManageLookups/DeleteLookupItemCommandHandler.cs
@@ -3,21 +3,24 @@
     using Entities.Lookups;
     using NPoco;
 
-    public class DeleteLookupSubtableDetailCommandHandler : ICommandHandler<DeleteLookupSubtableDetailCommand, bool>
+    public class DeleteLookupItemCommandHandler : ICommandHandler<DeleteLookupItemCommand, bool>
     {
         private readonly IDatabase _database;
 
-        public DeleteLookupSubtableDetailCommandHandler(IDatabase database)
+        public DeleteLookupItemCommandHandler(IDatabase database)
         {
             _database = database;
         }
-        public bool Handle(DeleteLookupSubtableDetailCommand message)
+        public bool Handle(DeleteLookupItemCommand message)
         {
-            var lookupType = LookupEntity.GetTypeFromName(message.Model.LookupType);
+            if (message.Id == 0)
+                return false;
+
+            var lookupType = LookupEntity.GetTypeFromName(message.LookupType);
             var singleByIdMethod = typeof (Database).GetMethod("SingleById");
             var genericSingleByIdMethod = singleByIdMethod.MakeGenericMethod(lookupType);
 
-            object[] args = {message.Model.Id};
+            object[] args = {message.Id};
 
             var result = genericSingleByIdMethod.Invoke(_database, args);
             _database.Delete(result);
